Implemented PlayComputerModal
